Renamed DeferredStream to BufferedSink and modled it more like react/stream's BufferedSink as suggested by @jsor at https://github.com/reactphp/http/pull/41#issuecomment-237160339
