+++ tests/StreamingBodyParser/BufferedSinkTest.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace React\Tests\Http;
+namespace React\Tests\Http\StreamingBodyParser;
 
 use Clue\React\Block;
 use React\EventLoop\Factory;
-use React\Http\DeferredStream;
 use React\Http\File;
+use React\Http\StreamingBodyParser\BufferedSink;
 use React\Http\StreamingBodyParser\NoBodyParser;
 use React\Http\Request;
 use React\Stream\ThroughStream;
-use React\Tests\Http\StreamingBodyParser\DummyParser;
+use React\Tests\Http\TestCase;
 
 class DeferredStreamTest extends TestCase
 {
     public function testDoneParser()
     {
         $parser = new NoBodyParser(new Request('get', 'http://example.com'));
-        $deferredStream = DeferredStream::create($parser);
+        $deferredStream = BufferedSink::createPromise($parser);
         $result = Block\await($deferredStream, Factory::create(), 10);
         $this->assertSame([
             'post' => [],
@@ -28,7 +28,7 @@ class DeferredStreamTest extends TestCase
     public function testDeferredStream()
     {
         $parser = new DummyParser(new Request('get', 'http://example.com'));
-        $deferredStream = DeferredStream::create($parser);
+        $deferredStream = BufferedSink::createPromise($parser);
         $parser->emit('post', ['foo', 'bar']);
         $parser->emit('post', ['array[]', 'foo']);
         $parser->emit('post', ['array[]', 'bar']);
@@ -72,12 +72,12 @@ class DeferredStreamTest extends TestCase
     public function testExtractPost()
     {
         $postFields = [];
-        DeferredStream::extractPost($postFields, 'dem', 'value');
-        DeferredStream::extractPost($postFields, 'dom[one][two][]', 'value_a');
-        DeferredStream::extractPost($postFields, 'dom[one][two][]', 'value_b');
-        DeferredStream::extractPost($postFields, 'dam[]', 'value_a');
-        DeferredStream::extractPost($postFields, 'dam[]', 'value_b');
-        DeferredStream::extractPost($postFields, 'dum[sum]', 'value');
+        BufferedSink::extractPost($postFields, 'dem', 'value');
+        BufferedSink::extractPost($postFields, 'dom[one][two][]', 'value_a');
+        BufferedSink::extractPost($postFields, 'dom[one][two][]', 'value_b');
+        BufferedSink::extractPost($postFields, 'dam[]', 'value_a');
+        BufferedSink::extractPost($postFields, 'dam[]', 'value_b');
+        BufferedSink::extractPost($postFields, 'dum[sum]', 'value');
         $this->assertSame([
             'dem' => 'value',
             'dom' => [
